Fix potential crashes because of JIT'ing bugs...
Presumably:

new QDir(...).entryList();

becomes inlined to:

       long id;
       {
           QDir x = new QDir(...);
           id = x.nativeId();
       } <- GC happens here

       QDir.__qt_entryList(id);

       Which means that the underlying QDir will have been destructed or
can be destructed while the entryList() function is being executed.
This is rather nasty, and definitly a bug in the JVM, but since its
present in 1.6.0 we have to work around it.

Instead of having the __qt_Xxx() native functions as static we make them
instance functions. That guarantees that there is a reference to the
instance during the function call.

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 290869]