p4i integration Integrate 303403 from 4.4 to main: If you tried to do something like
	new QFile(new QFileInfo("classpath:myfile").absoluteFilePath())

this would fail, because the class path engine failed to resolve 
paths given as "classpath:<location>#<path>" when the <location>
was a directory on disk. The reason for this is that the location
is given as an URL, and when we passed this to the File constructor
it would fail to parse it. Notably, we can unfortunately not use
new File(new URL(url).toURI()). Actually, it's discussable whether 
this is actually unfortunate, since the File constructor requires 
a hierarchical URI which means absolute paths have to start with 
a slash which means that we and our users would have to do a platform
specific check on Windows as to whether the path started with 
X: where X is a single letter, and then add a slash to the beginning
of this path. This is so horrible that I chose to strip away the 
"file:" scheme from the URL instead and check whether the remaining
path is a file or directory.

Specific case this fixes: e.g. new QIcon("classpath:mypixmap.png") 
       since this uses QFileInfo.absoluteFilePath() to get the absolute
path of the pixmap file.

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 303404]