p4i integration Integrate 305517 from 4.4 to main: Two things:
1. Refactored locate_vm() on Linux and Solaris, since
these two had lots of duplicated code. I've kept the one 
difference between the two functions in a #ifdef clause 
inside the new, common function instead.

2. Made locate_vm() look for the virtual machine in more
locations. It still uses the same default, but it now
also works if you set any of the supported environment 
variables to point to the JRE instead of the JVM. The
default on OpenSuSE for instance is that JAVA_HOME
points to /usr/lib/jvm/jre which means that adding
/jre/lib gets us nowhere. In the case when the vm is not
found in $THE_ENV/jre/lib we will now also try just
$THE_ENV/lib and work out of the box on OpenSuSE hopefully.

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 305518]