Problem: QAbstractTextDocumentLayout::registerHandler was using a pattern we can not support in Qt Jambi, as it depends on a user extending QObject and implementing the QTextObjectInterface interface. The latter is of not possible because we cannot create C++ objects with dynamically specified superclasses, and we obviously cannot support all permutations of multiple inheritance. So we needed some way of giving the equivalent/similar API in Java.
We chose to remove the QTextObjectInterface class from the type system, 
       and introduce a new class QtJambiTextObjectInterface which, on the C++
side inherits QObject and QTextObjectInterface. In Java the class will
be renamed to QTextObjectInterface, and you will have to extend it 
in order to make a handler for text objects. 

       A side effect of removing the C++ version of QTextObjectInterface
from the type system was that the handlerForObject() method disappeared 
from the API as the return type was now unknown. This has been injected
manually. I've also deprecated the registerHandler method which takes
a QObject and introduced a registerHandler() that takes a proper 
(java-) QTextObjectInterface instead. The original registerHandler is 
still called, and it throws an exception if the class passed in is wrong.

I've also added tests for this.

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 321624]