p4i integration Integrate 322336 from 4.5.0_01 to main: Make it possible to subclass QGraphicsScene while reimplementing itemChange() in QGraphicsItem.
In change 276875 I made a change to avoid crashes in QDynamicMetaObject. As the change description
says, this limits usage of QVariants. With QGraphicsItem::itemChange() we assume that any subclass of 
QGraphicsScene will become a QGraphicsScene* type in a QVariant, which does not work if we default
to JObjectWrapper for all user types. Result: Implementing itemChange() in a graphics item and 
subclassing QGraphicsScene will cause the item to not be added.

       Since we want to retain the solution in 276875 I have reverted behavior only for the QVariant case
and kept the fix for the QDynamicMetaObject case which should solve both issues.

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 322337]