p4i integration Integrate 322668 from 4.5.0_01 to main: Memory management patches...
The idea is that instead of using explicit reference counting	
on the JNI side, we let QObject's have a children array
which we populate as children are added and removed. This
makes it possible for us to use Java GC to clean up stuff.

The patch mostly works for the trivial test cases, but there
are the following problems:

 - QTreeView crashes on startup
   --> problem with out-of-index access in model?

 - "Drag and Drop" and "Draggable Text" crashes on startup
   --> looks problematic...
   --> related to qtjambi_setup_connections
   --> could it be that we have some out of sync stuff with link and
       ourselves where links are being leaked...

 - I18N is broken too...
   --> Crashes inside QTextLine...

 - Screenshot throws a nullpointer...

 - TextEdit Demo crashes on startup...
   --> NullPointer again...

The problem seems to be related to objects we believe have cpp owner
ship are collected behind our back...

[git-p4: depot-paths = "//depot/qtjambi/main/": change = 322669]