Make QAbstractXmlNodeModel an object type and add in missing methods
We can't remove methods in a patch release, so I've reimplemented some
methods that disappeared because QAbstractXmlNodeModel and
QSimpleXmlNodeModel were made object types. These methods would
previously not work and the types could not be instantiated, so the
impact of this is very low. I will vote to break bc for unimportant
cases in 4.6 and remove these deprecated methods.