Fix crash on shutdown in HelloGL
When disposed() is called, there is no guarantee that the C++ object
is valid, since disposed() might be called as a reaction to the object
being deleted from C++. In HelloGL, what happened was that we disposed
the top level widget, which deleted its children. The GLWidget
destructor deleted GL-specific data, then later on the QObject
destructor was called, which in turn deleted the user data for the
GLWidget, which in turn caused the disposed() method to be called.