Add some documentation for QtJambiObject.disposed()
It's possible to create a crash when reimplementing disposed(). If the
object is of a proper QObject subclass and is deleted from C++, then
its data will be deleted prior to the call to disposed() meaning that
you get a crash if you call methods on the object which accesses this
data. I have also made a task to consider ways of making this more
robust.