Avoid crashing when building meta object for classes with custom loader
When building the dynamic meta object, we need to resolve the classes
in signatures of slots and property accessors, e.g. to check their
inheritance to see if the classes belong to Qt and can easily be
converted to related Qt classes. When the classes cannot be resolved
on the class path (if they are loaded with a custom class loader),
the internalTypeName() method will throw an exception which is not
handled, and the resulting meta object becomes undefined. By handling
the exception and ignoring it, we avoid unwanted exception messages
on the console, and potential crashes.