Remove all generator.xmlmerge.* targets and associated conditions
Instead, use the generator.xmlmerge.check target to condition the execution of
generator.xmlmerge, in which we use the <xmlmerge> macro for all components.

The only difference to the previous behaviour is that all XSL transforms will be
run each time if only one file is changed, but the gain in clarity amply makes
for it.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>