Rerun generator when it has failed to run previously
I have encountered the case when running generator triggered a SIGSEGV.
Unfortunately, if I ran ant again to reproduce the problem, without an ant clean
previously, the generator.run target would not be run again.

The reason was that its condition to run at all only checked whether it was
_compiled_, not whether it was actually run.

Fix this by introducing a <condition> statement prior to the generator.run
statement that defines an appropriate "unless" property.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>