Add a basic .gitignore file
The build.xml's clean target is not complete, and there was no .gitignore at
all. As a result, git status shows a bucketload of files.

The .gitignore file created here just introduces the basic compilation
artifacts, *.o and *.class. This reduces the bucketload, but there is still a LOT
of work to do, especially since the output directory has the very bad idea to be
the same as the source directory, and qt-jambi generates Java and C++ source
files which we want to track (.gitignore) and not delete (build.xml's clean
target).

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>