Move the <dotimestamp> macro definition from build.xml to setenv.xml
The goal is to have setenv.xml download, extract and source ant-contrib tasks.
We want to be able to avoid the download and extract phases if they are
unnecessary, and this requires creating a timestamp file.

Until now, this was required only in build.xml. Handling ant-contrib, which is
done at the setup stage (therefore in setenv.xml), changes that. Therefore, move
the necessary include and macro definition from build.xml to setenv.xml

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>