Don't require ant-contrib.jar to be in the classpath anymore
This commit plugs all *.antcontrib tasks into the build system. This means that
one does not need to have ant-contrib downloaded separately. However, this
does mean that a network access is required for build.

It should be noted also that the clean target also removes all traces of the
previous downloads. Not ideal, but this stretches the need for a separate
distclean target.

Signed-off-by: Francis Galiegue <fgaliegue@gmail.com>