Designer/uic: Start on support for QIcon::fromTheme().
Schemas: Add "theme" attribute to <iconset> element of the UI format.

uic: Extend it to generate the code for icons with themes and/or resource
paths, add auto-test for it.

Designer: Add 'theme'-member to PropertySheetIconValue + convenience
functions including subproperty mask handling and simplify the existing
code. Add new class IconThemeEditor to be used in action editor dialog and
property editor. Theme names can now be specified in the action editor
dialog.

Reviewed-by: Jarek Kobus <jaroslaw.kobus@nokia.com>
Task-number: QTBUG-7777

Change-Id: I96983f35a2a6b9b5464ad9ce16bf04fc0610a598
Signed-off-by: Joni <joao267@gmail.com>
Reviewed-on: http://gerrit.smar.fi/308
Reviewed-by: Samu Voutilainen <smar@smar.fi>