HACK: This effectively removes the "const &" qualifiers from any QFlag<> enums.  QFlag is handled in a special way by QtJambi and the existing code always passes by reference by making a clone of the original object given.
It would have been nicer to find a spot higher up the generator APIs to
resolve this issue, however in my testing I could not find any spot that
could isolate this case.

The reason for this change as to allow the QScriptValue::PropertyFlag
to be implemented, since the QScriptValue::setProperty() methods
use the "const &" qualifier (the only part of the Qt API that does this).

The generated output has been checked that only the intended kind of
C++ CodeModel targets are edited in this way.