Move OSInfo out of package com.trolltech.qt.internal and into package com.trolltech.qt.osinfo
This is due to Java ClassLoader rules that by default a ClassLoader implementation
does not need to honour an exhaustive search of all CLASSPATH entries to find
a given resource that matches.  If it finds a JAR that contains at least one resource
in a given package then it can consider that JAR to be authoritative for that package
and thus it is possible for qtjambi-util.jar to mask classes in qtjambi.jar.

Practically a ClassLoader implementation may merge different resources for the same
package found in different JARs to form a unified view of the package.  However since
there is no guarantee of that and thus it is not a good default for QtJambi to maintain.