Code review: Fix unintended use of comma operator for a semi-colon.
It looks unintended as the indentation and lack of braces follows the code
indentation style.  If the comma operator was correct java references would
be leaked.  If the comma operator was correct the if no m_java_object existed
on function entry it would continue to not exist on exit (having been leaked).

Use of semi-colon instead makes complete sense given the context of the code.

-3 v +1 semi-colon wins.