QTextCodec is a kind of static global in Qt.  Meaning everytime you ask for it you get returned the same instance, that many users may share.  I think this is due to how they work in relation to plugin codecs.
However the typesystem XML correctly tries to assign ownership to cplusplus.

But generator.exe does not support that feature for a constructor.  The
index="-1" has a special meaning, to mean the return value and the return
value of a constructor is the instance being created.

The autotests have a subclass of QTextCodec which also needs to follow
the same rules.

The crashes that reliably occur are during application shutdown.  When NULLing
out the qtjambi_java_vm is moved to OnUnload.

I have attempted in the "#if 0" section to write a proper fix for this by detecting
that we are emitting code for a constructor and checking the ownership rules
however as I am pressed for time I have hardwired the 2 use cases by checking the
class names.

There is another inter-play with GlobalStaticDeleter actions performed on shutdown
and the QTJAMBI_DEBUG_TOOLS code, this only affects C++ public objects available
from API to QtJambi that are also involved with .dtor section of executables.

TODO: Implement the correct fix to support the XML configuration set for the constructor.