TestCase: Revamp TestMemoryManagement to make use of java.lang.ref.* to provide proof of lifecycle. This also has the effect of speeding up the unit testing, since we now have proof is passed without resorting to some kind of wall clock delays and some hope.
1 failing testcase remains:
com.trolltech.autotests.TestMemoryManagementNonPolymorphicObject#nativeDelete_NotCreatedInJava_CppOwnership()
 linkConstructedCount expected:<1> but was:<0>
 junit.framework.AssertionFailedError: linkConstructedCount expected:<1> but was:<0>
 at com.trolltech.autotests.TestMemoryManagement.test(TestMemoryManagement.java:232)
 at com.trolltech.autotests.TestMemoryManagement.createInNativeDisableGCAndDeleteInNative(TestMemoryManagement.java:503)
 at com.trolltech.autotests.TestMemoryManagement.nativeDelete_NotCreatedInJava_CppOwnership(TestMemoryManagement.java:480)

This does not appear to be the test cases fault but interaction between the
qtjambilink.cpp and the gUserObjectCache() in that the previous tests link
is not invalidated from the cache during setCppOwnership() and this failing
test cases happens to get the same malloc() memory to use for the next structure
but its still in the cache.  This hints at a caching/invalidation bug.