Qt4.8 support, initial update (just enough to get it to compile and unit tests to pass). Also tested against Fedora16 (which ships with 4.8.x on the system).
It was necessary to add support in generator to pass Option(SkipRemovedArguments) in respect of
constructor arguments, this is due to a change with the constructor of QCoreApplication
and QApplication.  Since I am in this area I exposed some of the additional constructors for
QApplication that have been talked about in the mailing-list over the past 6 months, possibly
allowing use of QtJambi in non-GUI applications.

TODO: I am sure there is a chunk of new 4.8 API that more attention in the typesystem to remove
the NativePointer from public visibility in method signatures.