TestCase: The last remaining 100% reliable JVM crash from test cases problem hass now been understood.  In effect the testcase was doing C++ equivalent of:
QWidget w = new QWidget();
QStackedLayout l = new QStackedLayout();
l.QLayout::addWidget(w);   // error line
delete l;
delete w;

The method is renamed to #addStackedWidget(QWidget) because of the use of 'final'
on the generated Java classes (which is also an incorrect design choice).  Strictly
speaking the full method signature is also changed (change of return type) as
well so even when 'final' keyword is removed, we should re-implement those method
to throw an Exception since it is invalid usage.