Ant introduce target "init.build".
Which ensures configuration for building/packaging
of QtJambi.  We keep "init" to be a more lightweight configuration needed to perform
non-building tasks such as project clean, unit test running, etc...
One reason for this change is a problem is the enviroment setup for unit testing is
as per the environment used for building QtJambi.  This is not the best scenario for
us.  We want the environment used for unit testing to be as close to a real users
deployment environment to help show up additional problems.
The full build environment is needed to build the unit tests target "tests.compile".
I would recommend you invoke ant separately with "tests.run" to run them to ensure the
minimal environment is then being used.