Modify Launcher main(String[]) to make it more robust to demonstrate QtJambi.
Due to the use of local hard references to QObject in the same method as the
initilize() execStatic() and shutdown() they do not get any opportunity to
be GCed until that method returns (i.e. after shutdown() has completed). This
has caused reliable crashing to occur on Linux64/JDK6 for example.