Update OSGi Manifest data based on realworld testing in an OSGi framework.
We remove Import/Export package versioning (since we do not properly
maintain a database to perform version number increments correctly).  We
use Require-Bundle to link the correct native implementation with the
main Java API bundle.  We add a Require/Provide-Capability which hopes to
replace the use of Require-Bundle in the future after that mechanism has
been proven effective.  Added the missing Export-Package for
com.trolltech.qt.test.