QSignalEmitterInternal reviewed: cross-thread emitting, auto prune dead connections.
Changed to allow cross-thread signal emitting.

Moved the check for QObject#signalsBlocked() inside the point where we
go from emitting/queuing to the delivery phase.  This method is not
thread-safe and in C++ it is only called from the receiver affinity thread
at the time of delivery.  In QtJambi it was being called at the time of
emitting which could be from the signal affinity thread or emitters thread.

This means it was not called safely anyway, when emitting from a thread
that was not the signal affinity thread it stopped the signal from being
delivered at all (it was not even given a chance to become a queued
signal).  Also by not letting it queue it does not give the signal a
chance to be delivered due to the QObject#blockSignals(boolean) state
changing being the time of emission and the time the message queue pump
gets to run again to deliver a queued signal.

Also when an exception occurs use a finally block to better indicate that
the restoration of currentSender needs to be done even under exception
conditions.

Also when the work was done to make connections to receiver be a
WeakReference by default it is possible for a null value to be returned
when the receiver has been GCed.  So the logic to automatically remove
dead connections did not account for that state as instanceof will always
return false if the object is null causing the dead connection to clog up
the Connections list.