Add the word/label "native-" into the JAR so we can call it the "native JAR".
This helps with documentation and understanding of the project, we can
in a platform independent way talk about "the native JAR" and users who
might not understand what that means will at least see a JAR with the word
native in it.  Sounds like a silly point to someone who understands the use
of the word "native" means; but it can help get people started building apps
with the project.

I am also adding the header X-QtJambi-SoName-Version the verdict is still
out on if it will help resolve some OSGi matter.