MetaObjectTools reflection bug IndexOutOfBoundsException on method set()
If you created a QObject subclass and trying to use QtJambi features on
it like signal/slots and you had a method called "public void set();" this
code would fail as it tried to grab the character after the "set" part at
the start of the method name since it did not check the string length for
one or more additional characters, thus causing IndexOutOfBoundsException.

Also had a quick look over the file for other similar looking issues and
made the Annotation / Qt Property handling a little more robust.