QtJambiInternal methods handling raw memory pointers use package visibility.
Even though the methods are inside an *.internal.* package and should not
be used they are still accessible from a SecurityManager point of view.
However I was unable to find a scenario that needed to use the methods so
by default the policy should be to use the lowest visibility possible on
methods working with raw memory pointers.