ThreadLocal usage robustness improvements.
Make the ThreadLocal variable 'final' as it has protected visibility but
should never be able to be changed.
Use a finally{} block to better indicate the manipulation and restoration
of the ThreadLocal value must always be done during error recovery.
Provide a mechanism to allow a fail-safe cleanup of thread local data
(ready for when Qt has QThread hook).