AbstractSignal.connectSignalMethod(Method, QObject, int) returns Void.
Many years ago this method did return 'boolean' much like the C++ API
does.  It looks like this JNI invocation was not updated at the time of
that change.

Also when looking a little over the history of this change the method now
throws an Exception when there is a problem, so we use this indicator to
return false in the function now.