Bug#141 signal connected to slot in the same object with the signal/slot named the same doesn't work
Update MetaData to revision 4, 5 or 6 based on Qt version.  The cause of the
problem was a non-backward compatible change in Qt made for 4.7.x series in
a patch qt.git commit-id afb4eee5.  The mechanism to filter by signal only
or slot only using the method flags field was removed.  This meant Qt
could see a slot method with the same signature as the signal.  This
pattern is used by QtJambi.

By upgrading the MetaData revision by workaround the matter since Qt will
use the new signalCount field of the QMetaObject descriptor.

This fixes the testcase TestSlotSameMethodName which is designed to
isolate this issue.

This has been unit tested on 4.6.4+ (branch:4.6) and v4.7.0, 4.7.4+
(branch:4.7) and v4.8.2.