Make SignalX final in qtjambi-designer.
This is a security concern as any code could re-assign the Signal
thus removing connections it did not want to keep, possibly causing
a bypass of some GUI control.  This may not be true for designer and
the cases changed here but it why default the advice should be to
make all SignalX final in all applications.

Change-Id: Idf64b9ad75670940586a5e957fa3e2b96475c5ac