TESTCASE: TestQSettings fix Java7 issues with reflection method order changes.
Java7 may return the test methods reflected in any order, some tests relied on
method ordering to be as per *.java file and therefore *.class order.  This
assumption has been addressed and removed.

Made the Serializable class static so it does not have a reference to its
enclosing class and thus will not try to serialize that too.  Which means
we can add fields to improve test and use clean settings file every run.

Change-Id: Ia3699be0313d8bbc22bc496730ffc11a7e0a3343