CODE-WALK qtjambi_core.cpp and qt_thread/QThread when reaping threads
Minor bugfix of adding a missing DeleteWeakGlobalRef() since it still
needs to be deallocated as the 'reference' itself is a hard global
reference, but the Java object the 'reference' points to does so with
a weak reference.

Another (simplified) way to think about it is that in JNI we obtain a
hard reference to a WeakReference object.  So we must deallocate the
JNI hard reference to allow that to be GCed.  In practise a JVM
provides optimization for this case so there is transparent referencing
by default.

Added comments and separate out some things into their own variable to
help debug printf and assertions.

Change-Id: If1fbddb60461c97c9bd4b0d167f6a1fe2edb9d71