registerSubObject debug accounting
This patch provides additional Q_ASSERT and checking for this part of
the API.  This should better document this part of the API by explaining
the contraints in which it operates.

It also resolves the C++ pointer aliases where a C++ type uses the same
memory pointer value for multiple views of the object (different
sub-types/interfaces/inheritance matters).

We also optimize a little when setting more than one pointer for a
specific instance when obtaining the hash lock.

Change-Id: I7ed88313eee90d5794ea95b6b42d5744e934a203