TESTCASE TestInjectedCode.java change to help debugging issues when these fails
An obversation with this test has been that it is particularly brittle compared
to other tests.  I believe this to be due to some kind of common error/problem
but it would be better to understand what it is and test for that specifically.
Maybe it is due to plugin loading, environment setup, Image handling maybe the
tests rely on a particular order they are run in.  Not sure at this time since
it is currently passing unit testing.

Change-Id: I4a83264c5f2934d216b360c0343c5e933658900d