PlatformJarTask: Disable libpng bundling for now
This will allow repository builds like lauchpad to compile successfully.
Fixes bug #313 http://redmine.smar.fi/issues/313 that was preventing 
successful builds in Ubuntu based distros.

Regarding libpng bundling, it is not required for repository builds, and 
as far as the website releases are concerned, we can just compile Qt 
with -qt-libpng flag to bundle it.

Change-Id: I508f581a28e00efe3b525b626f795f5b25b3675e
Reviewed-on: http://gerrit.smar.fi/1283
Reviewed-by: Samu Voutilainen <smar@smar.fi>